// models/enums.ts
export enum Sections {
  Projects = 'projects',
  Insights = 'insights'
};

export enum Categories {
  All = 'all',
  UxEngineering = 'uxEngineering',
  UxUiDesign = 'uxUiDesign',
  Design3D = '3dDesign',
  WebDevelopment = 'webDevelopment',
  MobileDevelopment = 'mobileDevelopment',
  CreativeConcepts = 'creativeConcepts',
  CaseStudies = 'caseStudies',
  IndependentResearch = 'independentResearch',
  DesktopDevelopment = 'desktopDevelopment',
};

export enum CareerProfiles {
  Designer,
  Engineer,
};

export enum Locales {
  French = 'fr',
  English = 'en'
};

export enum Pages {
  Home = 'home',
  About = 'about',
  Portfolio = 'portfolio',
  Services = 'services',
  Contact = 'contact',
  Story = 'story',
};

export enum Positions {
  Right = 'right',
  Left = 'left',
  LeftAndRight = 'leftRight',
  None = 'none',
  Up = 'Up',
  Down = 'Down',
};